Move SoundTranslator
