+++ proxy/src/main/java/org/dragonet/proxy/DragonProxy.java
@@ -30,8 +30,8 @@ import org.dragonet.proxy.configuration.ServerConfig;
 import org.dragonet.proxy.events.EventManager;
 import org.dragonet.proxy.network.RaknetInterface;
 import org.dragonet.proxy.network.SessionRegister;
+import org.dragonet.proxy.network.translator.SoundTranslator;
 import org.dragonet.proxy.utilities.ProxyLogger;
-import org.dragonet.proxy.utilities.SoundTranslator;
 import org.dragonet.proxy.utilities.MetricsManager;
 import org.dragonet.proxy.utilities.PluginManager;
 import org.dragonet.proxy.utilities.pingpassthrough.PingThread;
